> Refactor Code > Transfer from dedicated episode client and apolloclient to the initialized client > Transfer from dedicated episode mapper to CharacterMapper.kt > Code Rework/Reformat > Created GetAllEpisodeUseCase.kt and GetAllEp > Change of variable name in episodes data class
