+++ app/src/main/java/com/example/rickandmorty/api/ApolloModule.kt
@@ -2,12 +2,10 @@ package com.example.rickandmorty.api
 
 import com.apollographql.apollo3.ApolloClient
 import com.example.rickandmorty.data.ApolloCharacterClient
-import com.example.rickandmorty.data.ApolloEpisodeClient
-import com.example.rickandmorty.domain.EpisodeClient
+// import com.example.rickandmorty.data.ApolloEpisodeClient
 import com.example.rickandmorty.domain.character.GetCharacterUseCase
-import com.example.rickandmorty.domain.episodeusecase.GetEpisodeUseCase
-import com.example.rickandmorty.domain.episodeusecase.GetEpisodesUseCase
 import com.example.rickandmorty.domain.CharacterClient
+import com.example.rickandmorty.domain.episodeusecase.GetAllEpisodeUseCase
 import com.example.rickandmorty.domain.location.GetAllLocationUseCase
 import com.example.rickandmorty.domain.location.GetLocationDetailUseCase
 import dagger.Module
@@ -51,10 +49,16 @@ object ApolloModule {
         return GetCharacterUseCase(characterClient)
     }
 
+//    @Provides
+//    @Singleton
+//    fun provideGetEpisodeClient(apolloClient: ApolloClient): CharacterClient {
+//        return ApolloCharacterClient(apolloClient)
+//    }
+
     @Provides
     @Singleton
-    fun provideGetEpisodeClient(apolloClient: ApolloClient): EpisodeClient {
-        return ApolloEpisodeClient(apolloClient)
+    fun provideGetAllEpisodeUseCase(characterClient: CharacterClient): GetAllEpisodeUseCase {
+        return GetAllEpisodeUseCase(characterClient)
     }
 
     @Provides
@@ -63,17 +67,17 @@ object ApolloModule {
         return GetAllLocationUseCase(characterClient)
     }
 
-    @Provides
-    @Singleton
-    fun provideGetEpisodesUseCase(episodeClient: EpisodeClient): GetEpisodesUseCase {
-        return GetEpisodesUseCase(episodeClient)
-    }
+//    @Provides
+//    @Singleton
+//    fun provideGetEpisodesUseCase(characterClient: CharacterClient): GetEpisodesUseCase {
+//        return GetEpisodesUseCase(characterClient)
+//    }
 
-    @Provides
-    @Singleton
-    fun provideGetEpisodeUseCase(episodeClient: EpisodeClient): GetEpisodeUseCase {
-        return GetEpisodeUseCase(episodeClient)
-    }
+//    @Provides
+//    @Singleton
+//    fun provideGetEpisodeUseCase(characterClient: CharacterClient): GetEpisodeUseCase {
+//        return GetEpisodeUseCase(characterClient)
+//    }
 
     @Provides
     @Singleton
