+++ app/src/main/java/com/example/rickandmorty/data/ApolloCharacterClient.kt
@@ -1,15 +1,14 @@
 package com.example.rickandmorty.data
 
 import com.apollographql.apollo3.ApolloClient
-import com.example.AllLocationsQuery
-import com.example.CharactersQuery
-import com.example.LocationDetailQuery
+import com.example.*
 import com.example.rickandmorty.domain.location.Location
 import com.example.rickandmorty.domain.location.LocationDetail
 import com.example.rickandmorty.domain.character.Character
 import com.example.rickandmorty.domain.CharacterClient
+import com.example.rickandmorty.domain.DetailedEpisode
 import com.example.rickandmorty.domain.character.DetailedCharacter
-import com.example.SpecificCharacterQuery
+import com.example.rickandmorty.domain.Episodes
 
 class ApolloCharacterClient(private val apolloClient: ApolloClient) : CharacterClient {
     override suspend fun getCharacters(): List<Character> {
@@ -44,5 +43,25 @@ class ApolloCharacterClient(private val apolloClient: ApolloClient) : CharacterC
     override suspend fun getSingleCharacter(code: String): DetailedCharacter? {
         return apolloClient.query(SpecificCharacterQuery(code))
             .execute().data?.character?.toSpecificChar()
+    }
+
+    override suspend fun getEpisodes(): List<Episodes> {
+        return apolloClient
+            .query(GetEpisodesQuery())
+            .execute()
+            .data
+            ?.episodes
+            ?.results
+            ?.mapNotNull { it?.toEpisodes() }
+            ?: emptyList()
+    }
+
+    override suspend fun getEpisode(id: String): DetailedEpisode? {
+        return apolloClient
+            .query(GetEpisodeQuery(id))
+            .execute()
+            .data
+            ?.episode
+            ?.toDetailedEpisode()
     }
 }
