+++ app/src/main/java/com/example/rickandmorty/domain/CharacterClient.kt
@@ -14,4 +14,8 @@ interface CharacterClient {
     suspend fun getCharacters(): List<Character>
 
     suspend fun getSingleCharacter(code: String): DetailedCharacter?
+
+    suspend fun getEpisodes(): List<Episodes>
+
+    suspend fun getEpisode(id: String): DetailedEpisode?
 }
