+++ app/src/main/java/com/example/rickandmorty/domain/episodeusecase/GetAllEpisodeUseCase.kt
@@ -3,7 +3,9 @@ package com.example.rickandmorty.domain.episodeusecase
 import com.example.rickandmorty.domain.CharacterClient
 import com.example.rickandmorty.domain.Episodes
 
-class GetEpisodesUseCase(private val characterClient: CharacterClient) {
+class GetAllEpisodeUseCase(
+    private val characterClient: CharacterClient,
+) {
     suspend fun execute(): List<Episodes> {
         return characterClient
             .getEpisodes()
