+++ app/src/main/java/com/example/rickandmorty/domain/episodeusecase/GetEpisodeUseCase.kt
@@ -1,12 +1,12 @@
 package com.example.rickandmorty.domain.episodeusecase
 
+import com.example.rickandmorty.domain.CharacterClient
 import com.example.rickandmorty.domain.DetailedEpisode
-import com.example.rickandmorty.domain.EpisodeClient
 
 class GetEpisodeUseCase(
-    private val episodeClient: EpisodeClient,
+    private val characterClient: CharacterClient,
 ) {
     suspend fun execute(id: String): DetailedEpisode? {
-        return episodeClient.getEpisode(id)
+        return characterClient.getEpisode(id)
     }
 }
